requirements.txt
Django>=4.0
boto3
psycopg2-binary
dj-database-url
python-dotenv
gunicorn
whitenoise
